move-over: attempted early removal optimization doesn't work in all cases
An interrupting move may attempt to reuse an out-of-viewport sprite, and if we stop animating them early they will not be able to animate back in correctly. To re-attempt this we need a way to explicitly opt into a removed sprite being destroyed so there's no gap between "stop animating it" and "don't include it in the next transition".