Adding mark-sprite capability
This is a better alternative to `explode` as currently implement in liquid-fire. It lets you mark parts of the DOM that should be treated as separate sprites during animation.

Next step here is to improve the API for canceling out the motion of your parent in the cases where it's moving too, and improving the API for namespacing animators.