bugfix: restructure tasks in animated-each
There needs to be an overal "animating" task that is restartable, and the identity of that task must be passed to "willAnimate".

There also needs to be a task for tracking whether we are currently running the animation vs starting the animation, because we make different decisions in either case when we are interrupted by descendant animation.