Streamlined initial render handling
This switches us to a model where we always run transitions at initial render, but we (1) choose not to include the insertedSprites unless you explicit opt in, and (2) optimize away the transition if there are no sprites.

Together these still mean you probably don't get a transition at initial render in simple cases, but they make it easier to handle receivedSprites.