animated-beacon support
An animated-beacon exists so that other animators can match against it. It doesn't animating anything of its own, but it servers as a source and destination for other sprites.

A canonical use case for animated-beacon is to wrap it around a button that launches a modal, and then animate the modal so that it starts at the button.