Safer use of htmlSafe
The `htmlSafe` function can easily become a security footgun. Since the demo app exists to teach, I think it's important to exemplify the correct way to use it, even though there aren't actual security problems in the demo app since it has no authority and no private data.

The problem is that `htmlSafe` doesn't make a thing safe, it merely vouches that it's already safe.

So to use `htmlSafe` correctly, it should really be called by the same code that *makes* the string safe. In general it should be possible to look at a call to `htmlSafe` and immediately see *why* it's safe. The code that does the work of constructing the allegedly-safe string should be as close as possible to the call to `htmlSafe`.

The danger in keeping them apart is that a future refactor can easily violate the security assumptions and you can't see the problem just be examining the code that changed.