locking down ember-data version
As used by addon-docs. Otherwise when testing without yarn.lock, ember-cli-addon-docs brings in a version of ember-data too new to work on the older end of our tested range of ember versions.

Yet another example of what's bad about having addon-docs in the dummy app. The library needs to work on ember back to 2.12. But the docs app doesn't.