Fix for running in Fastboot
Fastboot does not have the concept of Element, CharacterData and
presumably DocumentType.  Inclusion of this polyfill thus makes
Fastboot crash early.

Implementation may well be changed to check for IE11 instead of
checking if the entities exist.  I opted for the latter as the
polyfill might support other cases now, or in the future.

Further, I based myself on fetching these entities from window as
TypeScript gave errors if I chose not to.