Fix types for addon/element-remove
The typescript file does not know of the definitions on window.  In
order to support these, we followed [1] resulting in an interface
addition to the Window interface.

As per that blog-post and the referred documentation at [2], the
interface declaration will be merged with the existing Window
declaration, rather than overwriting it.

[1] https://mariusschulz.com/blog/declaring-global-variables-in-typescript#augmenting-the-window-interface
[2] https://www.typescriptlang.org/docs/handbook/declaration-merging.html#merging-interfaces