feat: Update faker to use new fork.
[Marak Squires went rogue in
January](https://fakerjs.dev/about/announcements/2022-01-14.html#i-heard-something-happened-what-s-the-tldr)
and deleted Faker. Since then, a new group has taken over the library.
This PR updates the docs and test code to use the new library.

The main reason for doing the update is because as long as I remember,
the avatar images used in the docs for animating between routes haven't
worked. Now that I see this library is in v1, I figured it was time to
contribute and get working avatars to show.